+++ enemy.h
@@ -14,7 +14,7 @@ class Enemy: public Character {
         Hero *hero = nullptr;
         bool hasCompass = false;
         bool moveable;
-		static bool isHostile;
+	static bool isHostile;
         Floor *floor = nullptr;
     public:
         void setHero(Hero *hero);
@@ -23,9 +23,9 @@ class Enemy: public Character {
         char getType() const;
         void moveNext();
         void attack();
-		virtual void setHostile(bool isHostile) = 0;
-		virtual bool getHostile() const = 0;
-		virtual int dropGold() const = 0;
+	virtual void setHostile(bool isHostile) = 0;
+	virtual bool getHostile() const = 0;
+	virtual double dropGold() const = 0;
         virtual ~Enemy() = 0;
 };
