+++ merchant.h
@@ -8,7 +8,7 @@ class Merchant final: public Enemy {
         static bool isHostile;
     public:
         Merchant();
-        int dropGold() const override;
+        double dropGold() const override;
         void setHostile(bool isHostile) override ;
         bool getHostile() const override ;
         ~Merchant();
